Correct the context switch implementation
The context switch is the switch of context between kernel processes,
which is different from kernel entry (which is the switch between
user process and kernel process). Therefore, the stack point that
has to be stored will be `sp` instead of `psp`. Moreover, this adds
the `naked` attribute to the function to ensure the compiler doesn't
do anything unexpected, for example, loading the argument onto the
stack.